# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DriveLocal API
  version: 0.3.0
  description: |
    Backend for the single-driver ride-booking app serving the Kriel /
    Thubelihle area in ZAR. One owner-driver, booked by customers via SMS-OTP.

    Auth: request an OTP, then verify it to receive a `token` (JWT). Send the
    token as `Authorization: Bearer <token>` on role-gated routes. The owner
    driver's phone can also hit customer routes (to preview the app).
  contact:
    name: DriveLocal
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Point:
      type: object
      properties:
        lat: { type: number }
        lng: { type: number }
        address: { type: string, nullable: true }
        note: { type: string, nullable: true }
...
    Dispute:
      type: object
      properties:
        id: { type: string }
        tripId: { type: string }
        userId: { type: string }
        reason: { type: string, nullable: true }
        status: { type: string, enum: [open, resolved] }
        resolution: { type: string, nullable: true }
        createdAt: { type: string }
        resolvedAt: { type: string, nullable: true }
    Promo:
      type: object
      properties:
        id: { type: string }
        code: { type: string }
        discountPercent: { type: integer }
        maxUses: { type: integer }
        usedCount: { type: integer }
        validFrom: { type: string }
        validUntil: { type: string, nullable: true }
        active: { type: integer, enum: [0, 1] }
    CreatePromoBody:
      type: object
      required: [code, discountPercent, maxUses]
      properties:
        code: { type: string, example: FRIEND10 }
        discountPercent: { type: integer, minimum: 1, maximum: 100 }
        maxUses: { type: integer, minimum: 1 }
    UpsertSpotBody:
      type: object
      required: [name, lat, lng]
      properties:
        name: { type: string }
        category: { type: string, default: spot }
        address: { type: string, nullable: true }
        lat: { type: number }
        lng: { type: number }
        note: { type: string, nullable: true }
        sort: { type: integer, nullable: true }
        active: { type: integer, nullable: true }
    DriverAnalytics:
      type: object
      properties:
        dailyEarnings:
          type: array
          items:
            type: object
            properties:
              date: { type: string }
              label: { type: string }
              count: { type: integer }
              earnings: { type: number }
        busiestHours:
          type: array
          items:
            type: object
            properties:
              hour: { type: integer }
              count: { type: integer }
        acceptanceRate: { type: integer }
        cancelledByDriver: { type: integer }
        cancelledByCustomer: { type: integer }
        avgWaitMin: { type: integer }
        totalTrips: { type: integer }
        completedTrips: { type: integer }
paths:
...
  /api/driver/profile:
    put:
      tags: [Driver]
      summary: Update driver profile
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/me:
    get:
      tags: [Driver]
      summary: Driver profile
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/online:
    post:
      tags: [Driver]
      summary: Go online / offline
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isOnline]
              properties:
                isOnline: { type: boolean }
      responses:
        "200": { description: OK }
  /api/driver/location:
    post:
      tags: [Driver]
      summary: Publish driver location (REST; the app normally uses WebSocket)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                lat: { type: number }
                lng: { type: number }
                heading: { type: number, nullable: true }
                accuracy: { type: number, nullable: true }
      responses:
        "200": { description: OK }
  /api/driver/route:
    post:
      tags: [Driver]
      summary: Compute a navigation route
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                from: { $ref: '#/components/schemas/Point' }
                to: { $ref: '#/components/schemas/Point' }
      responses:
        "200": { description: OK }
  /api/driver/spots:
    get:
      tags: [Driver]
      summary: List pickup spots (incl. inactive)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    post:
      tags: [Driver]
      summary: Create a pickup spot
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertSpotBody' }
      responses:
        "200": { description: OK }
  /api/driver/spots/{id}:
    put:
      tags: [Driver]
      summary: Update a pickup spot
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
    delete:
      tags: [Driver]
      summary: Delete a pickup spot
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/active-trip:
    get:
      tags: [Driver]
      summary: Trip currently in progress
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/pending-trip:
    get:
      tags: [Driver]
      summary: The one outstanding requested trip
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/scheduled-trips:
    get:
      tags: [Driver]
      summary: Upcoming scheduled trips
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/trips:
    get:
      tags: [Driver]
      summary: Driver trip history
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/activate:
    post:
      tags: [Driver]
      summary: Move a scheduled trip into the live queue
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/accept:
    post:
      tags: [Driver]
      summary: Accept a requested trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/decline:
    post:
      tags: [Driver]
      summary: Decline a requested trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/arrive:
    post:
      tags: [Driver]
      summary: Arrived at pickup
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/start:
    post:
      tags: [Driver]
      summary: Start the trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/complete:
    post:
      tags: [Driver]
      summary: Complete the trip and settle payment
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                actualDistanceKm: { type: number }
                actualDurationMin: { type: number }
                tipAmount: { type: number, default: 0 }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/cancel:
    post:
      tags: [Driver]
      summary: Cancel a trip
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/trips/{id}/refund:
    post:
      tags: [Driver]
      summary: Refund a card payment
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: OK }
  /api/driver/earnings:
    get:
      tags: [Driver]
      summary: Earnings summary (today / week / all-time + tips + rating)
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/driver/analytics:
    get:
      tags: [Driver]
      summary: Analytics (daily earnings, busiest hours, acceptance rate, waits)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverAnalytics' }
  /api/driver/disputes:
    get:
      tags: [Driver]
      summary: List open fare disputes
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  disputes:
                    type: array
                    items: { $ref: '#/components/schemas/Dispute' }
  /api/driver/disputes/{id}/resolve:
    post:
      tags: [Driver]
      summary: Resolve a fare dispute
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                resolution: { type: string, nullable: true }
      responses:
        "200": { description: OK }
  /api/driver/promos:
    get:
      tags: [Driver]
      summary: List promo codes
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  promos:
                    type: array
                    items: { $ref: '#/components/schemas/Promo' }
    post:
      tags: [Driver]
      summary: Create a promo code
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePromoBody' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  promo: { $ref: '#/components/schemas/Promo' }
  /api/driver/settings:
    get:
      tags: [Driver]
      summary: Driver settings
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
    put:
      tags: [Driver]
      summary: Update driver settings
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: OK }
...